Kotlin Conversion cleanup & whitespace updates
