+++ build.gradle
@@ -1,12 +1,12 @@
 buildscript {
-    ext.kotlinVersion = '1.4.20'
+    ext.kotlinVersion = '1.4.31'
     repositories {
         jcenter()
         google()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.1'
+        classpath 'com.android.tools.build:gradle:4.1.2'
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
     }
@@ -20,5 +20,5 @@ allprojects {
 }
 
 ext {
-    exoPlayerVersion = "2.12.2"
+    exoPlayerVersion = "2.13.2"
 }
