+++ library/build.gradle
@@ -27,7 +27,7 @@ static def getLibraryInfo() {
 dependencies {
     api "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
 
-    implementation 'androidx.media:media:1.2.0'
+    implementation 'androidx.media:media:1.2.1'
     implementation 'androidx.appcompat:appcompat:1.2.0'
 
     // ExoPlayer
