+++ library/src/main/kotlin/com/devbrackets/android/exomedia/core/listener/CaptionListener.kt
@@ -23,5 +23,5 @@ import com.google.android.exoplayer2.text.Cue
  * A repeatListener for receiving notifications of timed text.
  */
 interface CaptionListener {
-    fun onCues(cues: List<Cue>)
+  fun onCues(cues: List<Cue>)
 }
