+++ library/src/main/kotlin/com/devbrackets/android/exomedia/core/listener/MetadataListener.kt
@@ -23,10 +23,10 @@ import com.google.android.exoplayer2.metadata.Metadata
  * A repeatListener for receiving ID3 metadata parsed from the media stream.
  */
 interface MetadataListener {
-    /**
-     * Called each time there is a metadata associated with current playback time.
-     *
-     * @param metadata The metadata.
-     */
-    fun onMetadata(metadata: Metadata)
+  /**
+   * Called each time there is a metadata associated with current playback time.
+   *
+   * @param metadata The metadata.
+   */
+  fun onMetadata(metadata: Metadata)
 }
