+++ library/src/main/kotlin/com/devbrackets/android/exomedia/core/video/ClearableSurface.kt
@@ -23,5 +23,5 @@ package com.devbrackets.android.exomedia.core.video
  * have their surfaces cleared.
  */
 interface ClearableSurface {
-    fun clearSurface()
+  fun clearSurface()
 }
