+++ library/src/main/kotlin/com/devbrackets/android/exomedia/core/video/scale/ScaleType.kt
@@ -21,25 +21,24 @@ package com.devbrackets.android.exomedia.core.video.scale
  * for each type
  */
 enum class ScaleType {
-    CENTER,
-    CENTER_CROP,
-    CENTER_INSIDE,
-    FIT_CENTER,
-    FIT_XY,
-    NONE;
+  CENTER,
+  CENTER_CROP,
+  CENTER_INSIDE,
+  FIT_CENTER,
+  FIT_XY,
+  NONE;
 
 
-    companion object {
-
-        /**
-         * Retrieves the [ScaleType] with the specified `ordinal`. If
-         * the ordinal is outside the allowed ordinals then [.NONE] will be returned
-         *
-         * @param ordinal The ordinal value for the [ScaleType] to retrieve
-         * @return The [ScaleType] associated with the `ordinal`
-         */
-        fun fromOrdinal(ordinal: Int): ScaleType {
-            return values().getOrElse(ordinal) { NONE }
-        }
+  companion object {
+    /**
+     * Retrieves the [ScaleType] with the specified `ordinal`. If
+     * the ordinal is outside the allowed ordinals then [.NONE] will be returned
+     *
+     * @param ordinal The ordinal value for the [ScaleType] to retrieve
+     * @return The [ScaleType] associated with the `ordinal`
+     */
+    fun fromOrdinal(ordinal: Int): ScaleType {
+      return values().getOrElse(ordinal) { NONE }
     }
+  }
 }
