+++ library/src/main/kotlin/com/devbrackets/android/exomedia/listener/OnCompletionListener.kt
@@ -22,8 +22,8 @@ package com.devbrackets.android.exomedia.listener
  * a media source has completed.
  */
 interface OnCompletionListener {
-    /**
-     * Called when the end of a media source is reached during playback.
-     */
-    fun onCompletion()
+  /**
+   * Called when the end of a media source is reached during playback.
+   */
+  fun onCompletion()
 }
