+++ library/src/main/kotlin/com/devbrackets/android/exomedia/listener/VideoControlsVisibilityListener.kt
@@ -18,16 +18,16 @@ package com.devbrackets.android.exomedia.listener
 
 /**
  * Interface definition for a callback to be invoked when the visibility
- * of the [com.devbrackets.android.exomedia.ui.widget.VideoControls] changes
+ * of the [com.devbrackets.android.exomedia.ui.widget.DefaultVideoControls] changes
  */
 interface VideoControlsVisibilityListener {
-    /**
-     * Occurs when the DefaultControls becomes visible
-     */
-    fun onControlsShown()
+  /**
+   * Occurs when the DefaultControls becomes visible
+   */
+  fun onControlsShown()
 
-    /**
-     * Occurs when the DefaultControls becomes hidden
-     */
-    fun onControlsHidden()
+  /**
+   * Occurs when the DefaultControls becomes hidden
+   */
+  fun onControlsHidden()
 }
