+++ library/src/main/kotlin/com/devbrackets/android/exomedia/ui/animation/TopViewHideShowAnimation.kt
@@ -25,48 +25,47 @@ import android.view.animation.AnimationSet
 import android.view.animation.TranslateAnimation
 
 /**
- * An animation used to slide [com.devbrackets.android.exomedia.ui.widget.VideoControls]
+ * An animation used to slide [com.devbrackets.android.exomedia.ui.widget.DefaultVideoControls]
  * in and out from the top of the screen when changing visibilities.
  */
 class TopViewHideShowAnimation(private val animationView: View, private val toVisible: Boolean, duration: Long) : AnimationSet(false) {
 
-    init {
+  init {
+    //Creates the Alpha animation for the transition
+    val startAlpha = (if (toVisible) 0 else 1).toFloat()
+    val endAlpha = (if (toVisible) 1 else 0).toFloat()
 
-        //Creates the Alpha animation for the transition
-        val startAlpha = (if (toVisible) 0 else 1).toFloat()
-        val endAlpha = (if (toVisible) 1 else 0).toFloat()
-
-        val alphaAnimation = AlphaAnimation(startAlpha, endAlpha)
-        alphaAnimation.duration = duration
-
-
-        //Creates the Translate animation for the transition
-        val startY = if (toVisible) -animationView.height else 0
-        val endY = if (toVisible) 0 else -animationView.height
-        val translateAnimation = TranslateAnimation(0f, 0f, startY.toFloat(), endY.toFloat())
-        translateAnimation.interpolator = if (toVisible) LinearOutSlowInInterpolator() else FastOutLinearInInterpolator()
-        translateAnimation.duration = duration
+    val alphaAnimation = AlphaAnimation(startAlpha, endAlpha)
+    alphaAnimation.duration = duration
 
 
-        //Adds the animations to the set
-        addAnimation(alphaAnimation)
-        addAnimation(translateAnimation)
+    //Creates the Translate animation for the transition
+    val startY = if (toVisible) -animationView.height else 0
+    val endY = if (toVisible) 0 else -animationView.height
+    val translateAnimation = TranslateAnimation(0f, 0f, startY.toFloat(), endY.toFloat())
+    translateAnimation.interpolator = if (toVisible) LinearOutSlowInInterpolator() else FastOutLinearInInterpolator()
+    translateAnimation.duration = duration
 
-        setAnimationListener(Listener())
+
+    //Adds the animations to the set
+    addAnimation(alphaAnimation)
+    addAnimation(translateAnimation)
+
+    setAnimationListener(Listener())
+  }
+
+  private inner class Listener : AnimationListener {
+
+    override fun onAnimationStart(animation: Animation) {
+      animationView.visibility = View.VISIBLE
     }
 
-    private inner class Listener : Animation.AnimationListener {
-
-        override fun onAnimationStart(animation: Animation) {
-            animationView.visibility = View.VISIBLE
-        }
-
-        override fun onAnimationEnd(animation: Animation) {
-            animationView.visibility = if (toVisible) View.VISIBLE else View.GONE
-        }
-
-        override fun onAnimationRepeat(animation: Animation) {
-            //Purposefully left blank
-        }
+    override fun onAnimationEnd(animation: Animation) {
+      animationView.visibility = if (toVisible) View.VISIBLE else View.GONE
     }
+
+    override fun onAnimationRepeat(animation: Animation) {
+      //Purposefully left blank
+    }
+  }
 }
