+++ library/src/main/kotlin/com/devbrackets/android/exomedia/util/LongExtensions.kt
@@ -27,15 +27,15 @@ import android.text.format.DateUtils
  * @return The human readable time
  */
 fun Long.millisToFormattedTimeString(): String {
-    if (this < 0) {
-        return "--:--"
-    }
-    val seconds = this % DateUtils.MINUTE_IN_MILLIS / DateUtils.SECOND_IN_MILLIS
-    val minutes = this % DateUtils.HOUR_IN_MILLIS / DateUtils.MINUTE_IN_MILLIS
-    val hours = this % DateUtils.DAY_IN_MILLIS / DateUtils.HOUR_IN_MILLIS
+  if (this < 0) {
+    return "--:--"
+  }
+  val seconds = this % DateUtils.MINUTE_IN_MILLIS / DateUtils.SECOND_IN_MILLIS
+  val minutes = this % DateUtils.HOUR_IN_MILLIS / DateUtils.MINUTE_IN_MILLIS
+  val hours = this % DateUtils.DAY_IN_MILLIS / DateUtils.HOUR_IN_MILLIS
 
-    return if (hours > 0) {
-        "%d:%02d:%02d".format(hours, minutes, seconds)
-    } else "%02d:%02d".format(minutes, seconds)
+  return if (hours > 0) {
+    "%d:%02d:%02d".format(hours, minutes, seconds)
+  } else "%02d:%02d".format(minutes, seconds)
 
 }
